// src/enums.ts
export enum AppConst {
    WIDTH = 1200,
    HEIGHT = 800,
    BACKGROUND_COLOR = 0x0000ff,
    PIER_WIDTH = 50,
    PIER_HEIGHT = 150,
    PIER_LINE = 7,
    PIER_COLOR = 0xFFFF00,
    SHIP_WIDTH = 80,
    SHIP_HEIGHT = 30,
    SHIP_LINE = 5,
    SHIP_COLOR_FULL = 0xFF0000,
    SHIP_COLOR_EMPTY = 0x00FF00,
    SPACING = 20,
    TIME_INTERVAL = 8000
}

export enum ShipType {
    Green = "green",
    Red = "red",
}